+++ test/test_simp.c
@@ -1428,6 +1428,22 @@ START_TEST(test_casegen)
 }
 END_TEST
 
+START_TEST(test_condexpr)
+{
+   set_standard(STD_19);
+
+   input_from_file(TESTDIR "/simp/condexpr.vhd");
+
+   tree_t a = parse_check_and_simplify(T_ENTITY, T_ARCH);
+
+   tree_t c1 = tree_decl(a, 0);
+   fail_unless(tree_kind(c1) == T_CONST_DECL);
+   fail_unless(tree_kind(tree_value(c1)) == T_LITERAL);
+
+   fail_if_errors();
+}
+END_TEST
+
 Suite *get_simp_tests(void)
 {
    Suite *s = suite_create("simplify");
@@ -1484,6 +1500,7 @@ Suite *get_simp_tests(void)
    tcase_add_test(tc_core, test_casefold1);
    tcase_add_test(tc_core, test_issue574);
    tcase_add_test(tc_core, test_casegen);
+   tcase_add_test(tc_core, test_condexpr);
    suite_add_tcase(s, tc_core);
 
    return s;